+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog #
 
+## 1.3.10 ##
+- Added the ability to choose whether or not resellers can view the API key
+- Some style edits to the module.zpm file (module page)
+- Changed the module numbering to allow 2 digit numbers in versions...
+
 ## 1.3.9 ##
 - Fixed some compatibility issues with PHP 5.3 (Thanks to charityz2 for reporting the issues and letting me debug using his VPS)
 - Removed some unneeded files from the dependencies of the Senitor API (documentation files, files for testing the dependencies)